requirements.txt
Pillow==10.3.0
torch==2.4.1
torchvision==0.19.1
numpy==1.26.4
gradio
transformers